# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Warehouse Management System API",
        "title": "Warehouse API",
        "contact": {},
        "version": "1.2"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
...
        "/v1/orders": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve orders with filtering, search, and cursor pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Get orders List",
                "parameters": [
                    {
                        "enum": [
                            "good",
                            "defective",
                            "unknown"
                        ],
                        "type": "string",
                        "description": "Filter by product status",
                        "name": "product_status",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "inbound",
                            "outbound"
                        ],
                        "type": "string",
                        "description": "Filter by type",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Created after date (YYYY-MM-DD)",
                        "name": "created_after",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Created before date (YYYY-MM-DD)",
                        "name": "created_before",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by product ID",
                        "name": "product_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by store ID",
                        "name": "store_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by department ID",
                        "name": "department_id",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "id",
                            "name",
                            "created_at"
                        ],
                        "type": "string",
                        "description": "Sort field",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "description": "Sort direction",
                        "name": "sort_order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Cursor for next page",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "description": "Number of items per page (max 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.orderListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create a order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Create a order",
                "parameters": [
                    {
                        "description": "Order object",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.OrderSummary"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/orders/export": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "Bearer": []
                    }
                ],
                "description": "Export filtered orders as an Excel (.xlsx) file. Supports the same query filters as the orders list endpoint.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Export orders to Excel",
                "parameters": [
                    {
                        "enum": [
                            "good",
                            "defective",
                            "unknown"
                        ],
                        "type": "string",
                        "description": "Filter by product status",
                        "name": "product_status",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "inbound",
                            "outbound"
                        ],
                        "type": "string",
                        "description": "Filter by order type",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Created after date (YYYY-MM-DD)",
                        "name": "created_after",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Created before date (YYYY-MM-DD)",
                        "name": "created_before",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by product ID",
                        "name": "product_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by store ID",
                        "name": "store_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by department ID",
                        "name": "department_id",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "id",
                            "created_at"
                        ],
                        "type": "string",
                        "description": "Sort field",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "description": "Sort direction",
                        "name": "sort_order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Excel file download",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/orders/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get a single order by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Get order by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.OrderSummary"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "Bearer": []
                    }
                ],
                "description": "Delete a order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Delete a order",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "Bearer": []
                    }
                ],
                "description": "Modify a order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Modify a order",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Order object with updated data",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "No Content - order successfully updated"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateOrderRequest": {
            "type": "object",
            "required": [
                "department_id",
                "expire_date",
                "price",
                "product_id",
                "product_status",
                "quantity",
                "store_id",
                "type"
            ],
            "properties": {
                "department_id": {
                    "type": "integer",
                    "example": 1
                },
                "description": {
                    "type": "string",
                    "example": "Stock intake"
                },
                "expire_date": {
                    "type": "string",
                    "format": "date",
                    "example": "2026-12-31"
                },
                "price": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 1000
                },
                "product_id": {
                    "type": "integer",
                    "example": 1
                },
                "product_status": {
                    "enum": [
                        "good",
                        "defective",
                        "unknown"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.ProductStatus"
                        }
                    ],
                    "example": "good"
                },
                "quantity": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 10
                },
                "store_id": {
                    "type": "integer",
                    "example": 1
                },
                "type": {
                    "enum": [
                        "inbound",
                        "outbound"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.OrderType"
                        }
                    ],
                    "example": "inbound"
                }
            }
        },
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "string"
                        },
                        "message": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "dto.OrderSummary": {
            "type": "object",
            "properties": {
                "department": {
                    "$ref": "#/definitions/dto.SimpleSummary"
                },
                "department_id": {
                    "type": "integer",
                    "example": 1
                },
                "description": {
                    "type": "string",
                    "example": "Stock intake"
                },
                "expire_date": {
                    "type": "string",
                    "format": "date",
                    "example": "2026-12-31"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "price": {
                    "type": "integer",
                    "example": 1000
                },
                "product": {
                    "$ref": "#/definitions/dto.SimpleSummary"
                },
                "product_id": {
                    "type": "integer",
                    "example": 1
                },
                "product_status": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.ProductStatus"
                        }
                    ],
                    "example": "good"
                },
                "quantity": {
                    "type": "integer",
                    "example": 10
                },
                "store": {
                    "$ref": "#/definitions/dto.SimpleSummary"
                },
                "store_id": {
                    "type": "integer",
                    "example": 1
                },
                "type": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.OrderType"
                        }
                    ],
                    "example": "inbound"
                }
            }
        },
...
        "dto.SimpleSummary": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "Product 1"
                }
            }
        },
...
        "dto.UpdateOrderRequest": {
            "type": "object",
            "properties": {
                "department_id": {
                    "type": "integer",
                    "example": 1
                },
                "description": {
                    "type": "string",
                    "example": "Updated note"
                },
                "expire_date": {
                    "type": "string",
                    "format": "date",
                    "example": "2026-12-31"
                },
                "price": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 500
                },
                "product_id": {
                    "type": "integer",
                    "example": 1
                },
                "product_status": {
                    "enum": [
                        "good",
                        "defective",
                        "unknown"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.ProductStatus"
                        }
                    ],
                    "example": "defective"
                },
                "quantity": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 5
                },
                "store_id": {
                    "type": "integer",
                    "example": 1
                },
                "type": {
                    "enum": [
                        "inbound",
                        "outbound"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.OrderType"
                        }
                    ],
                    "example": "outbound"
                }
            }
        },
...
        "dto.orderListResponse": {
            "type": "object",
            "properties": {
                "has_more": {
                    "type": "boolean",
                    "example": true
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.OrderSummary"
                    }
                },
                "limit": {
                    "type": "integer",
                    "example": 20
                },
                "next_cursor": {
                    "type": "string",
                    "example": ""
                }
            }
        },
...
        "models.OrderType": {
            "type": "string",
            "enum": [
                "inbound",
                "outbound"
            ],
            "x-enum-varnames": [
                "OrderTypeInbound",
                "OrderTypeOutbound"
            ]
        },
        "models.ProductStatus": {
            "type": "string",
            "enum": [
                "good",
                "defective",
                "unknown"
            ],
            "x-enum-varnames": [
                "ProductStatusGood",
                "ProductStatusDefective",
                "ProductStatusUnknown"
            ]
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Enter your API Key here",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "Bearer": {
            "description": "Enter JWT with Bearer prefix",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}